pyproject.toml
[project]
name = "tsmom"
version = "1.0.0"
description = "Production-grade Time-Series Momentum strategy system with multi-asset backtesting, regime detection, and interactive dashboard"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2",
    "numpy>=1.26",
    "scipy>=1.12",
    "yfinance>=0.2.36",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
    "streamlit>=1.32",
    "plotly>=5.19",
    "hmmlearn>=0.3",
    "tenacity>=8.2",
    "pyarrow>=15.0",
    "scikit-learn>=1.4",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.3", "mypy>=1.8"]

[project.scripts]
tsmom-dashboard = "src.dashboard.app:main"
